[MISC] move include/seqan3/{core,utility}/simd/all.hpp
