+++ test/snippet/test/simd_utility.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <seqan3/core/simd/all.hpp>
+#include <seqan3/utility/simd/all.hpp>
 #include <seqan3/test/simd_utility.hpp>
 
 TEST(simd, simd_eq)
