+++ test/snippet/utility/simd/simd_traits.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/core/simd/all.hpp>
+#include <seqan3/utility/simd/all.hpp>
 
 using uint16x8_t = seqan3::simd::simd_type_t<uint16_t, 8>;
