+++ test/snippet/utility/simd/simd_upcast.cpp
@@ -1,5 +1,5 @@
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/core/simd/all.hpp>
+#include <seqan3/utility/simd/all.hpp>
 
 using int16x8_t = seqan3::simd::simd_type_t<int16_t, 8>;
 using int32x4_t = seqan3::simd::simd_type_t<int32_t, 4>;
