+++ test/snippet/utility/simd/views/iota_simd_transform.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/core/simd/all.hpp>
+#include <seqan3/utility/simd/all.hpp>
 
 // The simd type with 8 unsigned shorts.
 using uint16x8_t = seqan3::simd_type_t<uint16_t, 8>;
